+++ book/src/assets/styles/page.module.scss.d.ts
@@ -1,13 +1,6 @@
-declare namespace PageModuleScssNamespace {
-  export interface IPageModuleScss {
-    github: string;
-    page: string;
-  }
+export interface IPageModuleScss {
+  'page': string;
+  'github': string;
 }
-
-declare const PageModuleScssModule: PageModuleScssNamespace.IPageModuleScss & {
-  /** WARNING: Only available when `css-loader` is used without `style-loader` or `mini-css-extract-plugin` */
-  locals: PageModuleScssNamespace.IPageModuleScss;
-};
-
-export = PageModuleScssModule;
+declare const styles: IPageModuleScss;
+export = styles;